refactor(agent): move onToolCall to its own module
