+++ app/(chat)/chat.tsx
@@ -12,8 +12,8 @@ import {
 import { Textarea } from "@/components/ui/textarea";
 import { ChatSchema, chatSchema } from "@/db/schema";
 import { useSyncedRef } from "@/hooks/use-sync-ref";
-import { getAccountBalance, matchInjectedAccount } from "@/lib/polkadot-api";
-import { AvailableApis, chainConfig, ChainConfig } from "@/papi-config";
+import { onChatToolCall } from "@/lib/ai";
+import { AvailableApis, ChainConfig } from "@/papi-config";
 import { ExtenstionContext } from "@/providers/extension-provider";
 import { useLightClientApi } from "@/providers/light-client-provider";
 import { useChat } from "@ai-sdk/react";
@@ -22,7 +22,6 @@ import {
   DefaultChatTransport,
   lastAssistantMessageIsCompleteWithToolCalls,
 } from "ai";
-import { SS58String } from "polkadot-api";
 import {
   InjectedExtension,
   InjectedPolkadotAccount,
@@ -70,173 +69,17 @@ export default function Chat() {
     }),
     async onToolCall({ toolCall }) {
       // client side tool execution
-      if (toolCall.toolName === "getBalances") {
-        const account = toolCall.input as { address: SS58String };
-        const balance = await getAccountBalance(
-          account.address,
-          apiRef,
-          activeChainRef,
-        );
-
-        void addToolResult({
-          tool: toolCall.toolName,
-          toolCallId: toolCall.toolCallId,
-          output: balance,
-        });
-      }
-
-      if (toolCall.toolName === "getConnectedAccounts") {
-        const accounts = connectedAccountsRef.current.map((account) => ({
-          name: account.name,
-          address: account.address,
-        }));
-
-        void addToolResult({
-          tool: toolCall.toolName,
-          toolCallId: toolCall.toolCallId,
-          output: JSON.stringify(accounts),
-        });
-      }
-
-      if (toolCall.toolName === "getActiveAccount") {
-        const active = selectedAccountRef.current;
-
-        void addToolResult({
-          tool: toolCall.toolName,
-          toolCallId: toolCall.toolCallId,
-          output: JSON.stringify({
-            name: active?.name,
-            address: active?.address,
-          }),
-        });
-      }
-
-      if (toolCall.toolName === "setActiveAccount") {
-        const account = toolCall.input as {
-          address: SS58String | undefined;
-          name: string;
-        };
-
-        const newAccount = matchInjectedAccount(account, selectedExtensionsRef);
-
-        if (newAccount) {
-          setSelectedAccountRef.current(newAccount.acc, newAccount.ext);
-          void addToolResult({
-            tool: toolCall.toolName,
-            toolCallId: toolCall.toolCallId,
-            output: {
-              success: true,
-              message: `Set active account to ${account.name}`,
-            },
-          });
-        } else {
-          void addToolResult({
-            tool: toolCall.toolName,
-            toolCallId: toolCall.toolCallId,
-            output: {
-              success: false,
-              message: `Account ${account.name} not found`,
-            },
-          });
-        }
-      }
-
-      if (toolCall.toolName === "getAvailableNetworks") {
-        const networks = chainConfig.map((chain) => chain.name);
-
-        void addToolResult({
-          tool: toolCall.toolName,
-          toolCallId: toolCall.toolCallId,
-          output: JSON.stringify(networks),
-        });
-      }
-
-      if (toolCall.toolName === "getActiveNetwork") {
-        const network = activeChainRef.current.name;
-
-        void addToolResult({
-          tool: toolCall.toolName,
-          toolCallId: toolCall.toolCallId,
-          output: network,
-        });
-      }
-
-      if (toolCall.toolName === "setActiveNetwork") {
-        const input = toolCall.input as {
-          chain: string;
-        };
-        const network = chainConfig.find(
-          (chain) => chain.name.toLowerCase() === input.chain.toLowerCase(),
-        );
-        if (input.chain === activeChainRef.current.name) {
-          void addToolResult({
-            tool: toolCall.toolName,
-            toolCallId: toolCall.toolCallId,
-            output: {
-              success: false,
-              message: `Network ${input.chain} is already active`,
-            },
-          });
-        }
-        if (network) {
-          void setActiveChainRef.current(network);
-          void addToolResult({
-            tool: toolCall.toolName,
-            toolCallId: toolCall.toolCallId,
-            output: {
-              success: true,
-              message: `Set active network to ${network.name}`,
-            },
-          });
-        } else {
-          void addToolResult({
-            tool: toolCall.toolName,
-            toolCallId: toolCall.toolCallId,
-            output: {
-              success: false,
-              message: `Network ${input.chain} not found`,
-            },
-          });
-        }
-      }
-
-      // if (toolCall.toolName === "getAvailableSystemChains") {
-      //   const system = chainConfig
-      //     .filter((chain) => {
-      //       return typeof chain.chainSpec.para_id === "number";
-      //     })
-      //     .map((chain) => {
-      //       return {
-      //         name: chain.name,
-      //         id: chain.chainSpec.para_id,
-      //       };
-      //     });
-
-      //   void addToolResult({
-      //     tool: toolCall.toolName,
-      //     toolCallId: toolCall.toolCallId,
-      //     output: JSON.stringify(system),
-      //   });
-      // }
-
-      // if (toolCall.toolName === "getAvailableRelayChains") {
-      //   const relay = chainConfig
-      //     .filter((chain) => {
-      //       return typeof chain.chainSpec.para_id === "undefined";
-      //     })
-      //     .map((chain) => {
-      //       return {
-      //         name: chain.name,
-      //         type: chain.chainSpec.para_id,
-      //       };
-      //     });
-
-      //   void addToolResult({
-      //     tool: toolCall.toolName,
-      //     toolCallId: toolCall.toolCallId,
-      //     output: JSON.stringify(relay),
-      //   });
-      // }
+      await onChatToolCall({
+        apiRef,
+        activeChainRef,
+        setActiveChainRef,
+        connectedAccountsRef,
+        selectedAccountRef,
+        setSelectedAccountRef,
+        selectedExtensionsRef,
+        toolCall,
+        addToolResult,
+      });
     },
   });
