+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "agent-dot",
-  "version": "1.7.8",
+  "version": "1.7.9",
   "private": true,
   "scripts": {
     "dev": "next dev --turbopack",
